+++ install-travisci.sh
@@ -34,8 +34,6 @@ pip install -r requirments.txt
 # Armadillo is licensed under the Apache License, Version 2.0
 git clone https://github.com/mir-am/armadillo-code.git temp
 
-
-
 # Building C++ extension module on Linux and OS X
 if [[ "$TRAVIS_OS_NAME" == "linux" ]]
 then
@@ -44,15 +42,16 @@ then
 # More info at https://bugs.python.org/issue25440
 # Following solution solves the problem on Travis CI
 # Get extension suffix for building Python extension module
-ext_suffix='./src/clippdcd'$(python3 -c $'from distutils.sysconfig import get_config_var; print(get_config_var(\'EXT_SUFFIX\'))')
+ext_suffix='./ltsvm/optimizer/clipdcd'$(python3 -c $'from distutils.sysconfig import get_config_var; print(get_config_var(\'EXT_SUFFIX\'))')
 
-g++ -O3 -Wall -shared -std=c++11 -fPIC `python3 -m pybind11 --includes` ./src/optimizer/pybind_clippdcd.cpp -o $ext_suffix -I ./temp/include -DARMA_DONT_USE_WRAPPER -lblas -llapack 
+g++ -O3 -Wall -shared -std=c++11 -fPIC `python3 -m pybind11 --includes` ./ltsvm/optimizer/pybind_clippdcd.cpp -o $ext_suffix -I ./temp/include -DARMA_DONT_USE_WRAPPER -lblas -llapack 
 
 elif [[ "$TRAVIS_OS_NAME" == "osx" ]]
 then
 echo $(python-config --extension-suffix)
-g++ -O3 -Wall -shared -std=c++11 -fPIC `python -m pybind11 --includes` ./src/optimizer/pybind_clippdcd.cpp -o ./src/clippdcd`python-config --extension-suffix` -I ./temp/include -DARMA_DONT_USE_WRAPPER -framework Accelerate
+g++ -O3 -Wall -shared -std=c++11 -fPIC `python -m pybind11 --includes` ./ltsvm/optimizer/pybind_clippdcd.cpp -o ./ltsvm/optimizer/clipdcd`python-config --extension-suffix` -I ./temp/include -DARMA_DONT_USE_WRAPPER -framework Accelerate
 fi
 
 # Creates result directory for saving unit test's output
 mkdir "result"
+