+++ ltsvm/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+LightTwinSVM Program - Simple and Fast
+Version: 0.3.0-alpha - 2018-06-03
+Developer: Mir, A. (mir-am@hotmail.com)
+License: GNU General Public License v3.0
+
+A place for package constants
+"""
+
+__version__ = '0.3.0-alpha'
+RELEASE_DATE = '2018-06-03'
+