+++ ltsvm/eval_classifier.py
@@ -13,12 +13,11 @@ train/test split, grid search and generating the detailed result.
 """
 
 
-from twinsvm import TSVM, MCTSVM
-from misc import progress_bar_gs, time_fmt
-from sklearn.model_selection import train_test_split
-from sklearn.model_selection import KFold
+from ltsvm.twinsvm import TSVM, MCTSVM, OVO_TSVM
+from ltsvm.misc import progress_bar_gs, time_fmt
+from sklearn.model_selection import train_test_split, KFold, ParameterGrid
 from sklearn.metrics import accuracy_score, recall_score, precision_score, f1_score
-from itertools import product
+#from itertools import product
 from datetime import datetime
 import numpy as np
 import pandas as pd
@@ -124,17 +123,20 @@ class Validator:
         self.validator = validator_type
         self.obj_TSVM = obj_tsvm
 
-    def cv_validator(self, c1=2**0, c2=2**0, gamma=2**0):
+    def cv_validator(self, dict_param):
 
         """
-        It applies cross validation on instance of TSVM classifier
+        It applies cross validation on instance of Binary TSVM classifier
+        Input:
+            dict_param: A dictionary of hyper-parameters (dict)
+            
         output:
             Evaluation metrics such as accuracy, precision, recall and F1 score
             for each class.
         """
         
         # Set parameters of TSVM classifer
-        self.obj_TSVM.set_parameter(c1, c2, gamma)
+        self.obj_TSVM.set_params(**dict_param)
         
         # K-Fold Cross validation, divide data into K subsets
         k_fold = KFold(self.validator[1])    
@@ -183,17 +185,17 @@ class Validator:
             fp = fp + accuracy_test[2]
             fn = fn + accuracy_test[3]
 
-        # m_a=0, m_r_p=1, m_p_p=2, m_f1_p=3, k=4, c1=5, c2=6, gamma=7,
-        # m_r_n=8, m_p_n=9, m_f1_n=10, tp=11, tn=12, fp=13, fn=14, iter=15    
-        return np.mean(mean_accuracy), np.std(mean_accuracy), [np.mean(mean_accuracy), \
-               np.std(mean_accuracy), np.mean(mean_recall_p), np.std(mean_recall_p), \
-               np.mean(mean_precision_p), np.std(mean_precision_p), np.mean(mean_f1_p), \
-               np.std(mean_f1_p), np.mean(mean_recall_n), np.std(mean_recall_n), \
-               np.mean(mean_precision_n), np.std(mean_precision_n), np.mean(mean_f1_n), \
-               np.std(mean_f1_n), tp, tn, fp, fn, c1, c2, gamma if self.obj_TSVM.kernel_t == 'RBF' else '']
+        return np.mean(mean_accuracy), np.std(mean_accuracy), {**{'accuracy': np.mean(mean_accuracy),
+                      'acc_std': np.std(mean_accuracy),'recall_p': np.mean(mean_recall_p),
+                      'r_p_std': np.std(mean_recall_p), 'precision_p': np.mean(mean_precision_p),
+                      'p_p_std': np.std(mean_precision_p), 'f1_p': np.mean(mean_f1_p),
+                      'f1_p_std': np.std(mean_f1_p), 'recall_n': np.mean(mean_recall_n),
+                      'r_n_std': np.std(mean_recall_n), 'precision_n': np.mean(mean_precision_n),
+                      'p_n_std': np.std(mean_precision_n), 'f1_n': np.mean(mean_f1_n),
+                      'f1_n_std': np.std(mean_f1_n), 'tp': tp, 'tn': tn, 'fp': fp,
+                      'fn': fn}, **dict_param}
 
-
-    def split_tt_validator(self, c1=2**0, c2=2**0, gamma=2**0):
+    def split_tt_validator(self, dict_param):
         
         """
         It trains TwinSVM classifier on random training set and tests the classifier
@@ -204,7 +206,7 @@ class Validator:
         """
 
         # Set parameters of TSVM classifer
-        self.obj_TSVM.set_parameter(c1, c2, gamma)
+        self.obj_TSVM.set_params(**dict_param)
 
         X_train, X_test, y_train, y_test = train_test_split(self.train_data, \
                                            self.labels_data, test_size=self.validator[1], \
@@ -219,18 +221,20 @@ class Validator:
         f1_n = eval_metrics(y_test, output)
 
        # m_a=0, m_r_p=1, m_p_p=2, m_f1_p=3, k=4, c1=5, c2=6, gamma=7,
-       # m_r_n=8, m_p_n=9, m_f1_n=10, tp=11, tn=12, fp=13, fn=14,   
-        return accuracy, 0.0, [accuracy, recall_p, precision_p, f1_p, recall_n, \
-               precision_n, f1_n, tp, tn, fp, fn, c1, c2, gamma if self.obj_TSVM.kernel_t == 'RBF' else '']
+       # m_r_n=8, m_p_n=9, m_f1_n=10, tp=11, tn=12, fp=13, fn=14,
+        return accuracy, 0.0, {**{'accuracy': accuracy, 'recall_p': recall_p,
+               'precision_p': precision_p, 'f1_p': f1_p, 'recall_n': recall_n,
+               'precision_n': precision_n, 'f1_n': f1_n, 'tp': tp, 'tn': tn,
+               'fp': fp, 'fn': fn}, **dict_param}
 
-    def cv_validator_mc(self, c=2**0, gamma=2**0):
+    def cv_validator_mc(self, dict_param):
 
         """
         It applies cross validation on instance of multiclass TSVM classifier
         """
 
         # Set parameters of multiclass TSVM classifer
-        self.obj_TSVM.set_parameter(c, gamma)
+        self.obj_TSVM.set_params(**dict_param)
 
         # K-Fold Cross validation, divide data into K subsets
         k_fold = KFold(self.validator[1])    
@@ -262,10 +266,12 @@ class Validator:
             mean_precision.append(precision_score(y_test, output, average='micro') * 100)
             mean_f1.append(f1_score(y_test, output, average='micro') * 100)
 
-        return np.mean(mean_accuracy), np.std(mean_accuracy), [np.mean(mean_accuracy), \
-               np.std(mean_accuracy), np.mean(mean_recall), np.std(mean_recall), \
-               np.mean(mean_precision), np.std(mean_precision), np.mean(mean_f1), \
-               np.std(mean_f1), c, gamma if self.obj_TSVM.kernel_t == 'RBF' else '']
+        return np.mean(mean_accuracy), np.std(mean_accuracy), {**{'accuracy':
+               np.mean(mean_accuracy), 'acc_std': np.std(mean_accuracy),
+               'micro_recall': np.mean(mean_recall), 'm_rec_std': np.std(mean_recall),
+               'micro_precision': np.mean(mean_precision), 'm_prec_std':
+               np.std(mean_precision), 'mirco_f1': np.mean(mean_f1), 'm_f1_std':
+               np.std(mean_f1)}, **dict_param}
 
     def choose_validator(self):
 
@@ -283,7 +289,7 @@ class Validator:
 
                 return self.split_tt_validator
 
-        elif isinstance(self.obj_TSVM, MCTSVM):  # Multi-class TSVM
+        else:  # Multi-class TSVM
 
             if self.validator[0] == 'CV':
 
@@ -303,19 +309,22 @@ def search_space(kernel_type, class_type, c_l_bound, c_u_bound, rbf_lbound, \
         return search elements for grid search (List)
     """
 
-    c_range = [2 ** i for i in np.arange(c_l_bound, c_u_bound + 1, step, dtype=np.float)]
+    c_range = [2**i for i in np.arange(c_l_bound, c_u_bound+1, step,
+                                         dtype=np.float)]
+    
+    gamma_range = [2**i for i in np.arange(rbf_lbound, rbf_ubound+1, step,
+                   dtype=np.float)] if kernel_type == 'RBF' else [1]
+    
+    if class_type == 'binary' or class_type == 'ovo':
+        
+        param_grid = ParameterGrid({'C1': c_range, 'C2': c_range,
+                                    'gamma': gamma_range})
 
-    if class_type == 'binary':
+    elif class_type == 'ova':
+        
+        param_grid = ParameterGrid({'C': c_range, 'gamma': gamma_range})
 
-        return list(product(*[c_range, c_range, ])) if kernel_type == 'linear' else \
-               list(product(*[c_range, c_range, [2 ** i for i in np.arange(rbf_lbound, \
-                    rbf_ubound + 1, step, dtype=np.float)]]))
-
-    elif class_type == 'multiclass':
-
-        return list(product(*[c_range, ])) if kernel_type == 'linear' else \
-               list(product(*[c_range, [2 ** i for i in np.arange(rbf_lbound, \
-                    rbf_ubound + 1, step, dtype=np.float)]]))
+    return list(param_grid)
 
 
 def grid_search(search_space, func_validator):
@@ -360,9 +369,12 @@ def grid_search(search_space, func_validator):
             #start_time = time.time()
 
             # Save result after each run
-            acc, acc_std, result = func_validator(*element)
+            acc, acc_std, result = func_validator(element)
 
             #end = time.time()
+            
+            # For debugging purpose
+            #print('Acc: %.2f+-%.2f | params: %s' % (acc, acc_std, str(result)))
 
             result_list.append(result)
 
@@ -405,24 +417,27 @@ def save_result(file_name, validator_obj, gs_result, output_path):
 
     column_names = {'binary': {'CV': ['accuracy', 'acc_std', 'recall_p', 'r_p_std', 'precision_p', 'p_p_std', \
                            'f1_p', 'f1_p_std', 'recall_n', 'r_n_std', 'precision_n', 'p_n_std', 'f1_n',\
-                           'f1_n_std', 'tp', 'tn', 'fp', 'fn', 'c1', 'c2','gamma'],
+                           'f1_n_std', 'tp', 'tn', 'fp', 'fn'],  #'c1', 'c2','gamma'],
                     't_t_split': ['accuracy', 'recall_p', 'precision_p', 'f1_p', 'recall_n', 'precision_n', \
-                                  'f1_n', 'tp', 'tn', 'fp', 'fn', 'c1', 'c2','gamma']},
+                                  'f1_n', 'tp', 'tn', 'fp', 'fn']},  #, 'c1', 'c2','gamma']},
                     'multiclass':{'CV': ['accuracy', 'acc_std', 'micro_recall', 'm_rec_std', 'micro_precision', \
-                                         'm_prec_std', 'mirco_f1', 'm_f1_std', 'C', 'gamma']}}
+                                         'm_prec_std', 'mirco_f1', 'm_f1_std']}} #'C', 'gamma']
 
     # (Name of validator, validator's attribute) - ('CV', 5-folds)
     validator_type, validator_attr = validator_obj.validator              
 
-    output_file = os.path.join(output_path, "TSVM_%s_%s_%s_%s.xlsx") % (validator_obj.obj_TSVM.kernel_t, \
-                  "%d-F-CV" % validator_attr if validator_type == 'CV' else 'Tr%d-Te%d' % \
+    output_file = os.path.join(output_path, "%s_%s_%s_%s_%s.xlsx") % (validator_obj.obj_TSVM.cls_name,
+                              validator_obj.obj_TSVM.kernel, "%d-F-CV" %
+                              validator_attr if validator_type == 'CV' else 'Tr%d-Te%d' % \
                   ((1.0 - validator_attr) * 100, validator_attr * 100),
                   file_name, datetime.now().strftime('%Y-%m-%d %H-%M'))
 
     excel_file = pd.ExcelWriter(output_file, engine='xlsxwriter')
-
-    result_frame = pd.DataFrame(gs_result, columns=column_names['binary' if \
-                   isinstance(validator_obj.obj_TSVM, TSVM) else 'multiclass'][validator_type]) 
+    
+    # columns=column_names['binary' if \isinstance(validator_obj.obj_TSVM, TSVM) else 'multiclass'][validator_type]
+    result_frame = pd.DataFrame(gs_result, columns=column_names['binary' if 
+                                isinstance(validator_obj.obj_TSVM, TSVM) else
+                                'multiclass'][validator_type] + validator_obj.obj_TSVM.get_params_names()) 
 
     result_frame.to_excel(excel_file, sheet_name='Sheet1', index=False)
 
@@ -443,8 +458,12 @@ def initializer(user_input_obj):
 
         tsvm_obj = TSVM(user_input_obj.kernel_type, user_input_obj.rect_kernel)
 
-    elif user_input_obj.class_type == 'multiclass':
+    elif user_input_obj.class_type == 'ovo':
 
+        tsvm_obj = OVO_TSVM(user_input_obj.kernel_type)
+        
+    elif user_input_obj.class_type == 'ova':
+        
         tsvm_obj = MCTSVM(user_input_obj.kernel_type)
 
     validate = Validator(user_input_obj.X_train, user_input_obj.y_train, \
@@ -455,8 +474,8 @@ def initializer(user_input_obj):
                       user_input_obj.upper_b_u)
 
     # Dispaly headers
-    print("TSVM-%s    Dataset: %s    Total Search Elements: %d" % (user_input_obj.kernel_type, \
-          user_input_obj.filename, len(search_elements)))
+    print("%s-%s    Dataset: %s    Total Search Elements: %d" % (tsvm_obj.cls_name,
+          user_input_obj.kernel_type, user_input_obj.filename, len(search_elements)))
 
     result = grid_search(search_elements, validate.choose_validator())