+++ ltsvm/twinsvm.py
@@ -23,46 +23,41 @@ from sklearn.utils.multiclass import check_classification_targets
 from sklearn.utils.validation import check_X_y, check_is_fitted, check_array
 from sklearn.utils import column_or_1d
 import numpy as np
-# ClippDCD optimizer is an extension module which is implemented in C++
-import clippdcd
+
+# ClipDCD optimizer is an extension module which is implemented in C++
+from ltsvm.optimizer import clipdcd
 
 
-class TSVM:
+class TSVM(BaseEstimator):
 
-    def __init__(self, kernel_type='linear', rect_kernel=1, c1=2**0, c2=2**0, \
+    def __init__(self, kernel='linear', rect_kernel=1, C1=2**0, C2=2**0, \
                  gamma=2**0):
 
         """
-            Input:
+        Input:
             Kernel_type: 1->Linear, 2->RBF(Gaussion)
             c1, c2: Penalty parameters
             gamma: RBF function parameter
         """
 
-        self.C1 = c1
-        self.C2 = c2
-        self.u = gamma
-        self.kernel_t = kernel_type
-        self.rectangular_size = rect_kernel
+        self.C1 = C1
+        self.C2 = C2
+        self.gamma = gamma
+        self.kernel = kernel
+        self.rect_kernel = rect_kernel
         self.mat_C_t = None
-
+        self.cls_name = 'TSVM'
+        
         # Two hyperplanes attributes
         self.w1, self.b1, self.w2, self.b2 = None, None, None, None
 
-    def set_parameter(self, c1=2**0, c2=2**0, gamma=2**0):
-
+    def get_params_names(self):
+        
         """
-        It changes the parametes for TSVM classifier.
-        DO NOT USE THIS METHOD AFTER INSTANTIATION OF TSVM CLASS!
-        THIS METHOD CREATED ONLY FOR Validator CLASS.
-        Input:
-            c1, c2: Penalty parameters
-            gamma: RBF function parameter
+        It returns the names of hyper-parameters of this classifier.
         """
-
-        self.C1 = c1
-        self.C2 = c2
-        self.u = gamma
+        
+        return ['C1', 'C2'] if self.kernel == 'linear' else ['C1', 'C2', 'gamma']
 
     def fit(self, X_train, y_train):
 
@@ -87,21 +82,21 @@ class TSVM:
         mat_e1 = np.ones((mat_A.shape[0], 1))
         mat_e2 = np.ones((mat_B.shape[0], 1))
 
-        if self.kernel_t == 'linear':  # Linear kernel
+        if self.kernel == 'linear':  # Linear kernel
             
             mat_H = np.column_stack((mat_A, mat_e1))
             mat_G = np.column_stack((mat_B, mat_e2))
 
-        elif self.kernel_t == 'RBF': # Non-linear 
+        elif self.kernel == 'RBF': # Non-linear 
 
             # class 1 & class -1
             mat_C = np.row_stack((mat_A, mat_B))
 
-            self.mat_C_t = np.transpose(mat_C)[:, :int(mat_C.shape[0] * self.rectangular_size)]
+            self.mat_C_t = np.transpose(mat_C)[:, :int(mat_C.shape[0] * self.rect_kernel)]
 
-            mat_H = np.column_stack((rbf_kernel(mat_A, self.mat_C_t, self.u), mat_e1))
+            mat_H = np.column_stack((rbf_kernel(mat_A, self.mat_C_t, self.gamma), mat_e1))
 
-            mat_G = np.column_stack((rbf_kernel(mat_B, self.mat_C_t, self.u), mat_e2))
+            mat_G = np.column_stack((rbf_kernel(mat_B, self.mat_C_t, self.gamma), mat_e2))
 
 
         mat_H_t = np.transpose(mat_H)
@@ -119,9 +114,9 @@ class TSVM:
         # Wolfe dual problem of class -1
         mat_dual2 = np.dot(np.dot(mat_H, mat_G_G), mat_H_t)
 
-        # Obtaining Lagrane multipliers using ClippDCD optimizer
-        alpha_d1 = np.array(clippdcd.clippDCD_optimizer(mat_dual1, self.C1)).reshape(mat_dual1.shape[0], 1)
-        alpha_d2 = np.array(clippdcd.clippDCD_optimizer(mat_dual2, self.C2)).reshape(mat_dual2.shape[0], 1)
+        # Obtaining Lagrange multipliers using ClipDCD optimizer
+        alpha_d1 = np.array(clipdcd.clippDCD_optimizer(mat_dual1, self.C1)).reshape(mat_dual1.shape[0], 1)
+        alpha_d2 = np.array(clipdcd.clippDCD_optimizer(mat_dual2, self.C2)).reshape(mat_dual2.shape[0], 1)
 
         # Obtain hyperplanes
         hyper_p_1 = -1 * np.dot(np.dot(mat_H_H, mat_G_t), alpha_d1)
@@ -149,14 +144,14 @@ class TSVM:
         prepen_distance = np.zeros((X_test.shape[0], 2))
 
         kernel_f = {'linear': lambda i: X_test[i, :] , 'RBF': lambda i: rbf_kernel(X_test[i, :], \
-                    self.mat_C_t, self.u)}
+                    self.mat_C_t, self.gamma)}
 
         for i in range(X_test.shape[0]):
 
             # Prependicular distance of data pint i from hyperplanes
-            prepen_distance[i, 1] = np.abs(np.dot(kernel_f[self.kernel_t](i), self.w1) + self.b1)
+            prepen_distance[i, 1] = np.abs(np.dot(kernel_f[self.kernel](i), self.w1) + self.b1)
 
-            prepen_distance[i, 0] = np.abs(np.dot(kernel_f[self.kernel_t](i), self.w2) + self.b2)
+            prepen_distance[i, 0] = np.abs(np.dot(kernel_f[self.kernel](i), self.w2) + self.b2)
 
         # Assign data points to class +1 or -1 based on distance from hyperplanes
         output = 2 * np.argmin(prepen_distance, axis=1) - 1
@@ -186,14 +181,14 @@ class HyperPlane:
         self.b = None  # Bias term
 
 
-class MCTSVM:
+class MCTSVM(BaseEstimator):
 
     """
     Multi Class Twin Support Vector Machine
     One-vs-All Scheme
     """
 
-    def __init__(self, kernel_type='linear', c=2**0, gamma=2**0):
+    def __init__(self, kernel='linear', C=2**0, gamma=2**0):
 
         """
         Input:
@@ -202,31 +197,27 @@ class MCTSVM:
             gamma: parameter of RBF function
         """
 
-        self.kernel_t = kernel_type
-        self.C = c
-        self.y = gamma
+        self.kernel = kernel
+        self.C = C
+        self.gamma = gamma
         self.classfiers = {}  # Classifiers
         self.mat_D_t = []  # For non-linear MCTSVM
+        self.cls_name = 'TSVM_OVA'
 
-    def set_parameter(self, c=2**0, gamma=2**0):
-
+    def get_params_names(self):
+        
         """
-        It changes the parametes for multiclass TSVM classifier.
-        DO NOT USE THIS METHOD AFTER INSTANTIATION OF MCTSVM CLASS!
-        THIS METHOD CREATED ONLY FOR Validator CLASS.
-        Input:
-            c: Penalty parameters
-            gamma: RBF function parameter
+        It returns the names of hyper-parameters of this classifier.
         """
-
-        self.C = c
-        self.y = gamma
+        
+        return ['C'] if self.kernel == 'linear' else ['C', 'gamma']
 
     def fit(self, X_train, y_train):
 
         """
-        X_train: Training samples
-        y_train: Lables of training samples
+        Input:
+            X_train: Training samples
+            y_train: Lables of training samples
         """
 
         num_classes = np.unique(y_train)
@@ -244,19 +235,19 @@ class MCTSVM:
             mat_e1_i = np.ones((mat_X_i.shape[0], 1))
             mat_e2_i = np.ones((mat_Y_i.shape[0], 1))
 
-            if self.kernel_t == 'linear':
+            if self.kernel == 'linear':
                 
                 mat_A_i = np.column_stack((mat_X_i, mat_e1_i))
                 mat_B_i = np.column_stack((mat_Y_i, mat_e2_i))
 
-            elif self.kernel_t == 'RBF':
+            elif self.kernel == 'RBF':
 
                 mat_D = np.row_stack((mat_X_i, mat_Y_i))
 
                 self.mat_D_t.append(np.transpose(mat_D))
 
-                mat_A_i = np.column_stack((rbf_kernel(mat_X_i, self.mat_D_t[idx], self.y), mat_e1_i))
-                mat_B_i = np.column_stack((rbf_kernel(mat_Y_i, self.mat_D_t[idx], self.y), mat_e2_i))
+                mat_A_i = np.column_stack((rbf_kernel(mat_X_i, self.mat_D_t[idx], self.gamma), mat_e1_i))
+                mat_B_i = np.column_stack((rbf_kernel(mat_Y_i, self.mat_D_t[idx], self.gamma), mat_e2_i))
 
             mat_A_i_t = np.transpose(mat_A_i)
             mat_B_i_t = np.transpose(mat_B_i)
@@ -271,7 +262,7 @@ class MCTSVM:
             mat_dual_i = np.dot(np.dot(mat_B_i, mat_A_A), mat_B_i_t)
     
             # Obtaining Lagrange multipliers using ClippDCD optimizer
-            alpha_i = np.array(clippdcd.clippDCD_optimizer(mat_dual_i, self.C)).reshape(mat_dual_i.shape[0], 1)
+            alpha_i = np.array(clipdcd.clippDCD_optimizer(mat_dual_i, self.C)).reshape(mat_dual_i.shape[0], 1)
     
             hyperplane_i = np.dot(np.dot(mat_A_A, mat_B_i_t), alpha_i)
     
@@ -294,13 +285,13 @@ class MCTSVM:
         prepen_dist = np.zeros((X_test.shape[0], len(self.classfiers.keys())))
 
         kernel_f = {'linear': lambda i, j: X_test[i, :] , 'RBF': lambda i, j: rbf_kernel(X_test[i, :], \
-                    self.mat_D_t[j], self.y)}
+                    self.mat_D_t[j], self.gamma)}
 
         for i in range(X_test.shape[0]):
 
             for idx, j in enumerate(self.classfiers.keys()):
 
-                prepen_dist[i, idx] = np.abs(np.dot(kernel_f[self.kernel_t](i, idx), \
+                prepen_dist[i, idx] = np.abs(np.dot(kernel_f[self.kernel](i, idx), \
                            self.classfiers[j].w) + self.classfiers[j].b) / np.linalg.norm(self.classfiers[j].w)
 
         output = np.argmin(prepen_dist, axis=1) + 1
@@ -317,7 +308,7 @@ class OVO_TSVM(BaseEstimator, ClassifierMixin):
     such as cross_val_score and GridSearchCV can be used for OVO_TSVM
     """    
     
-    def __init__(self, kernel='linear', C_1=1, C_2=1, gamma=1):
+    def __init__(self, kernel='linear', C1=1, C2=1, gamma=1):
         
         """
         Parameters:
@@ -330,9 +321,18 @@ class OVO_TSVM(BaseEstimator, ClassifierMixin):
         """
         
         self.kernel = kernel
-        self.C_1 = C_1
-        self.C_2 = C_2
+        self.C1 = C1
+        self.C2 = C2
         self.gamma = gamma
+        self.cls_name = 'TSVM_OVO'
+        
+    def get_params_names(self):
+        
+        """
+        It returns the names of hyper-parameters of this classifier.
+        """
+        
+        return ['C1', 'C2'] if self.kernel == 'linear' else ['C1', 'C2', 'gamma']
          
     def _validate_targets(self, y):
         
@@ -404,7 +404,7 @@ class OVO_TSVM(BaseEstimator, ClassifierMixin):
                 sub_prob_y_i_j[sub_prob_y_i_j == j] = -1
                 sub_prob_y_i_j[sub_prob_y_i_j == i] = 1
                 
-                self.bin_TSVM_models_[p] = TSVM(self.kernel, 1, self.C_1, self.C_2, \
+                self.bin_TSVM_models_[p] = TSVM(self.kernel, 1, self.C1, self.C2, \
                                self.gamma)
                 
                 self.bin_TSVM_models_[p].fit(sub_prob_X_i_j, sub_prob_y_i_j)
@@ -465,37 +465,42 @@ class OVO_TSVM(BaseEstimator, ClassifierMixin):
     
 if __name__ == '__main__':
     
-    from dataproc import read_data
-    from sklearn.metrics import accuracy_score
-    from sklearn.model_selection import train_test_split
+#    from ltsvm.dataproc import read_data
+#    from sklearn.metrics import accuracy_score
+#    from sklearn.model_selection import train_test_split
     from sklearn.utils.estimator_checks import check_estimator
-    from sklearn.model_selection import cross_val_score, GridSearchCV
-    import time
+#    from sklearn.model_selection import cross_val_score, GridSearchCV
+#    import time
+#    
+    check_estimator(OVO_TSVM)
 
     
-    train_data, labels, data_name = read_data('/home/mir/mir-projects/Mir-Repo/dataset/mc-data/wine.csv')
-    
-    X_train, X_test, y_train, y_test = train_test_split(train_data, labels,
-                                                        test_size=0.30, random_state=42)
-    
-    
-#    param = {'C_1': [float(2**i) for i in range(-5, 6)],
-#             'C_2': [float(2**i) for i in range(-5, 6)]}
-    
-    start_t = time.time()
+#    train_data, labels, data_name = read_data('/home/mir/mir-projects/Mir-Repo/mc-data/wine.csv')
 #    
-    ovo_tsvm_model = OVO_TSVM('linear')
-    
-    #cv = cross_val_score(ovo_tsvm_model, train_data, labels, cv=10)
-    
-#    result = GridSearchCV(ovo_tsvm_model, param, cv=10, n_jobs=4, refit=False, verbose=1)
-#    result.fit(train_data, labels)
-    
+#    X_train, X_test, y_train, y_test = train_test_split(train_data, labels,
+#                                                        test_size=0.30, random_state=42)
 #    
-    ovo_tsvm_model.fit(X_train, y_train)
-    
-    pred = ovo_tsvm_model.predict(X_test)
 #    
-    print("Finished: %.2f ms" % ((time.time() - start_t) * 1000))
+##    param = {'C_1': [float(2**i) for i in range(-5, 6)],
+##             'C_2': [float(2**i) for i in range(-5, 6)]}
 #    
-    print("Accuracy: %.2f" % (accuracy_score(y_test, pred) * 100))
+#    start_t = time.time()
+##    
+#    ovo_tsvm_model = MCTSVM()
+#    ovo_tsvm_model.set_params(**{'C': 4, 'gamma': 0.1})
+#    print(ovo_tsvm_model.get_params())
+#    
+#    #cv = cross_val_score(ovo_tsvm_model, train_data, labels, cv=10)
+#    
+##    result = GridSearchCV(ovo_tsvm_model, param, cv=10, n_jobs=4, refit=False, verbose=1)
+##    result.fit(train_data, labels)
+#    
+#    print(X_train.shape)
+##    
+#    ovo_tsvm_model.fit(X_train, y_train)
+#    
+#    pred = ovo_tsvm_model.predict(X_test)
+##    
+#    print("Finished: %.2f ms" % ((time.time() - start_t) * 1000))
+##    
+#    print("Accuracy: %.2f" % (accuracy_score(y_test, pred) * 100))